Integration to staging

+++ scripts/remove-key.sh
@@ -1,3 +1,5 @@
 #!/bin/sh
-security delete-keychain ios-build.keychain
-rm -f "~/Library/MobileDevice/Provisioning Profiles/$PROFILE_NAME.mobileprovision"
+if test "$TRAVIS_BRANCH" = 'staging'; then
+	security delete-keychain ios-build.keychain
+	rm -f "~/Library/MobileDevice/Provisioning Profiles/$PROFILE_NAME.mobileprovision"
+fi
